[core] Refactor database related classes
Gives more specific names to Module (DatabaseModule) and Query (DatabaseQuery), and renames LibraryDatabase to a more suitable TrackDatabase
